% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helper_function.R
\name{.calculate_columns_means}
\alias{.calculate_columns_means}
\title{calculate list of columns mean of target by column value}
\usage{
.calculate_columns_means(
  mean_target_colnames,
  target_colname,
  data,
  min_samples_leaf = 0,
  smoothing = 1
)
}
\arguments{
\item{mean_target_colnames}{list of column names that needs mean target encoding}

\item{target_colname}{target column name, should be numeric}

\item{data}{dataset used to calculate mean target, usually is the training dataset}

\item{min_samples_leaf}{minimum quantity per column value we consider needs to be calculate mean.}

\item{smoothing}{used to adjust column value mean target from grand target value, default value 1 is to ignore grand target value}
}
\value{
a data.table contains mean target per column value for list of columns
}
\description{
Output a data.table that contains list of columns mean target by column value
}
\keyword{internal}
