% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helper_function.R
\name{.calculate_single_colum_mean}
\alias{.calculate_single_colum_mean}
\title{calculate single column mean of target by column value}
\usage{
.calculate_single_colum_mean(
  mean_target_colname,
  target_colname,
  data,
  min_samples_leaf = 0,
  smoothing = 1
)
}
\arguments{
\item{mean_target_colname}{single column name that needs mean target encoding}

\item{target_colname}{target column name, should be numeric}

\item{data}{dataset used to calculate mean target, usually is the training dataset}

\item{min_samples_leaf}{minimum quantity per column value we consider needs to be calculate mean.
below two settings introduced here https://www.kaggle.com/ogrellier/python-target-encoding-for-categorical-features}

\item{smoothing}{used to adjust column value mean target from grand target value, default value 1 is to ignore grand target value}
}
\value{
a data.table contains mean target per column value
}
\description{
Output a data.table that contains single column mean target by column value
}
\keyword{internal}
