% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_mean_target_apply.R
\name{ml_mean_target_apply}
\alias{ml_mean_target_apply}
\title{Apply mean target encoding}
\usage{
ml_mean_target_apply(
  calculate_columns_mean_table,
  applying_dataset,
  impute_missing = T
)
}
\arguments{
\item{calculate_columns_mean_table}{mean target encoding mapping table outputed from function ml_mean_target_encoder}

\item{applying_dataset}{new dataset that you wish to apply same mean target encoding}

\item{impute_missing}{determine if we want to impute missing when new dataste levels could not be found in mapping table, defualt is to impute}
}
\value{
dataset that has been applied mean target encoding
}
\description{
This function is used to apply mean target encoding mapping table that was output from function ml_mean_target_encoder
}
\seealso{
\link{ml_mean_target_encoder}
}
